+++ examples/tts/conf/de/fastpitch_align_44100.yaml
@@ -67,7 +67,7 @@ model:
     punct_post_process: true
 
   text_tokenizer:
-    _target_: nemo.collections.tts.torch.tts_tokenizers.GermanCharsTokenizer
+    _target_: nemo.collections.common.tokenizers.text_to_speech.tts_tokenizers.GermanCharsTokenizer
     punct: true
     apostrophe: true
     pad_with_space: true
