+++ examples/tts/conf/mixer-tts-x.yaml
@@ -70,7 +70,7 @@ model:
     punct_post_process: true
 
   text_tokenizer:
-    _target_: nemo.collections.tts.torch.tts_tokenizers.EnglishCharsTokenizer
+    _target_: nemo.collections.tts.common.tokenizers.text_to_speech.EnglishCharsTokenizer
     punct: true
     apostrophe: true
     pad_with_space: true
