+++ nemo/collections/tts/torch/tts_dataset.yaml
@@ -30,7 +30,7 @@ tts_dataset:
     punct_post_process: True
 
   text_tokenizer:
-    _target_: nemo.collections.tts.torch.tts_tokenizers.EnglishPhonemesTokenizer
+    _target_: nemo.collections.common.tokenizers.text_to_speech.tts_tokenizers.EnglishPhonemesTokenizer
     punct: True
     stresses: True
     chars: True
@@ -41,6 +41,6 @@ tts_dataset:
     add_blank_at: null
     pad_with_space: True
     g2p:
-      _target_: nemo.collections.tts.torch.g2ps.EnglishG2p
+      _target_: nemo.collections.common.tokenizers.text_to_speech.g2ps.EnglishG2p
       phoneme_dict: "scripts/tts_dataset_files/cmudict-0.7b_nv22.07"
       heteronyms: "scripts/tts_dataset_files/heteronyms-030921"
