+++ scripts/dataset_processing/tts/hui_acg/ds_conf/ds_for_fastpitch_align.yaml
@@ -38,7 +38,7 @@ dataset:
     punct_post_process: true
 
   text_tokenizer:
-    _target_: nemo.collections.tts.torch.tts_tokenizers.GermanCharsTokenizer
+    _target_: nemo.collections.common.tokenizers.text_to_speech.tts_tokenizers.GermanCharsTokenizer
     punct: true
     apostrophe: true
     pad_with_space: true
