+++ tests/collections/tts/test_torch_tts.py
@@ -19,11 +19,11 @@ from pathlib import Path
 import numpy as np
 import pytest
 import torch
+from nemo_text_processing.g2p.modules import EnglishG2p
 
+from nemo.collections.common.tokenizers.text_to_speech.tts_tokenizers import EnglishPhonemesTokenizer
 from nemo.collections.tts.torch.data import TTSDataset
-from nemo.collections.tts.torch.g2ps import EnglishG2p
 from nemo.collections.tts.torch.helpers import get_base_dir
-from nemo.collections.tts.torch.tts_tokenizers import EnglishPhonemesTokenizer
 
 
 class TestTTSDataset:
